calculator_1

+++ 0x02-python-import_modules/calculator_1.py
@@ -41,4 +41,4 @@ def div(a, b):
     Returns:
         The return value. a / b
     """
-    return int(a / b)
+    return (a / b)
